Moved hmr reload fetch into the RootProvider
Created HmrContext so can set the shell root from the RootProvider. It works much better in RootProvider because can access the stateNavigator and context. Creating a new one was clumsy plus could get the wrong url because the current href won't necessarily match the state context (history action none, for example)
