+++ NavigationReact/sample/rsc-parcel/src/HmrContext.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export default createContext<any>(null);
